1.133 LeerCsv arreglado y refactor de base
