+++ src/jandas/base/data/Fila.java
@@ -4,10 +4,30 @@ import jandas.base.etiquetas.Etiqueta;
 
 import java.util.List;
 
-public interface Fila {
+public class Fila {
 
-    Etiqueta getEtiquetaFila();
-    List<Etiqueta> getEtiquetasColumnas();
-    List<Celda<?>> getCeldasFila();
-    String toString();
+    private Etiqueta etiquetaFila;
+    private List<Celda<?>> celdas;
+    private List<Etiqueta> etiquetaColumnas;
+
+    public Fila(Etiqueta etiquetaFila, List<Celda<?>> celdas, List<Etiqueta> etiquetaColumnas) {
+        this.etiquetaFila = etiquetaFila;
+        this.celdas = celdas;
+        this.etiquetaColumnas = etiquetaColumnas;
+    }
+
+    @Override
+    public String toString() {
+        String salida = "*" + etiquetaFila + "*" + " | ";
+        for (Celda<?> celda : celdas) {
+            salida += celda + " | ";
+        }
+        return salida;
+    }
+
+    public List<Etiqueta> getEtiquetasColumnas() {return etiquetaColumnas;}
+
+    public Etiqueta getEtiquetaFila() {return etiquetaFila;}
+
+    public List<Celda<?>> getCeldasFila() {return celdas;}
 }
