+++ src/jandas/io/EscribirArchivo.java
@@ -5,7 +5,4 @@ import jandas.base.data.Tabla;
 public interface EscribirArchivo  < O> {
 
     void escribir(Tabla tabla, String rutaArchivo);
-
-    void escribir(Tabla tabla, O config, String rutaArchivo);
-
 }
