+++ src/jandas/io/csv/EscribirCsv.java
@@ -3,14 +3,8 @@ package jandas.io.csv;
 import jandas.base.data.Tabla;
 import jandas.io.EscribirArchivo;
 
-public class EscribirCsv implements EscribirArchivo<CsvConfig> {
-    @Override
-    public void escribir(Tabla tabla, String rutaArchivo) {
-
-    }
+public class EscribirCsv implements EscribirArchivo {
 
     @Override
-    public void escribir(Tabla tabla, CsvConfig config, String rutaArchivo) {
-
-    }
+    public void escribir(Tabla tabla, String rutaArchivo) {}
 }
