+++ src/jandas/visualizacion/FormatoTabla.java
@@ -2,8 +2,8 @@ package jandas.visualizacion;
 
 import jandas.base.data.Celda;
 import jandas.base.data.Columna;
-import jandas.base.data.Tabla;
 import jandas.base.data.Fila;
+import jandas.base.data.Tabla;
 import jandas.base.etiquetas.Etiqueta;
 
 import java.util.ArrayList;
