+++ src/tests/VisualizacionDemo.java
@@ -1,10 +1,7 @@
 package tests;
 
 import jandas.base.data.Celda;
-import jandas.base.data.CeldaGenerica;
 import jandas.base.data.Tabla;
-import jandas.base.data.TablaGenerica;
-import jandas.base.etiquetas.Etiqueta;
 import jandas.base.etiquetas.EtiquetaInt;
 import jandas.base.etiquetas.EtiquetaString;
 import jandas.io.csv.CsvConfig;
@@ -13,7 +10,6 @@ import jandas.visualizacion.VConfig;
 import jandas.visualizacion.VConsola;
 
 import java.util.Arrays;
-import java.util.List;
 
 /**
  * Demo class showing how to use the visualization package.
@@ -22,14 +18,14 @@ public class VisualizacionDemo {
     
     public static void main(String[] args) {
         // Creo dataframe
-        Tabla df = new TablaGenerica();
+        Tabla df = new Tabla();
         
         // Agrego columnas usando los 3 metodos
 
         // metodo 1
         df.agregarColumna(new EtiquetaString("Nombre"),
                 String.class,
-                Arrays.asList("Juan", "María", "Pedro", "Ana", "Luis"));
+                Arrays.asList("Juan", "María", "Pedro" , "Ana", "Luis"));
         
         df.agregarColumna(new EtiquetaInt(1),
                 Integer.class,
@@ -43,10 +39,10 @@ public class VisualizacionDemo {
                 Boolean.class,
                 Arrays.asList(true, false, true, true, false));
 
-        Celda<String> c1 = new CeldaGenerica<>("ramon gutierrez");
-        Celda<Integer> c2 = new CeldaGenerica<>(4);
-        Celda<Double> c3 = new CeldaGenerica<>(1.5);
-        Celda<Boolean> c4 = new CeldaGenerica<>(true);
+        Celda<String> c1 = new Celda<>("ramon gutierrez");
+        Celda<Integer> c2 = new Celda<>(4);
+        Celda<Double> c3 = new Celda<>(1.5);
+        Celda<Boolean> c4 = new Celda<>(true);
 
         df.agregarFila(new EtiquetaString("fila 6"),
                 Arrays.asList(c1,c2,c3,c4));
@@ -78,22 +74,6 @@ public class VisualizacionDemo {
         System.out.println("Visualización con objeto de configuración personalizado:");
         visualizadorPersonalizado.visualizarConConfig(df, config);
 
-        Tabla m = new TablaGenerica();
-
-        Object[][] matriz = {
-                {"Juan", 25, "Argentina"},
-                {"Ana", 30, "Chile"},
-                {"Luis", 28, "Uruguay"}
-        };
-
-        List<Etiqueta> etiquetas = List.of(
-                new EtiquetaString("Nombre"),
-                new EtiquetaString("Edad"),
-                new EtiquetaString("País")
-        );
-
-        m.cargarDesdeMatriz(matriz, etiquetas);
-
 
         // iniciañoizo lector CSV
         LeerCsv lectorCsv = new LeerCsv();
@@ -109,6 +89,4 @@ public class VisualizacionDemo {
         Tabla df3 = (new LeerCsv()).leer("df/flights_september.csv", new CsvConfig(",", true));
         visualizador.visualizar(df3);
     }
-
-
 }
